requirements.txt
python-telegram-bot
groq
sqlalchemy
python-dotenv
pandas
openpyxl
matplotlib
sqlalchemy-utils
httpx
